Password reset cycle complete and stylized
- When user does not remember the password, he can
click on a link in page. He is taken to email
input.
- Fetch GET request is send to server. Server
creates a hash from SECRET and salt and stores
it with userID and timestamp in a database.
- Nodemailer sends email with link and UTM hash
to user.
- User click on link, that opens the reset
password inputs.
- Fetch request with hash and new password
body is send.
- Server finds object with same hash. If exists,
checks if timestamp is less than 15 minutes.
If true, creates new password has from password
and salt and it updates user object.
- User can now login with a new password.